# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dogonomics API for stock data, news, and sentiment. Swagger UI is available at /swagger/index.html",
        "title": "Dogonomics API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/chart/{symbol}": {
            "get": {
                "description": "Returns historical price data for a symbol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charts"
                ],
                "summary": "Get chart data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol (e.g., AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Days of history (max 365)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/finnews/{symbol}": {
            "get": {
                "description": "Returns recent news for a symbol (from EODHD)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get latest news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol (e.g., AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/sentAnalysis.NewsItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/finnewsBert/{symbol}": {
            "get": {
                "description": "Fetches news for a symbol and returns items with BERT sentiment and aggregate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sentiment"
                ],
                "summary": "Get news with BERT sentiment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol (e.g., AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.NewsSentimentBERTResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns API health information and data sources",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            }
        },
        "/profile/{symbol}": {
            "get": {
                "description": "Returns company profile for a symbol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Get company profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol (e.g., AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quote/{symbol}": {
            "get": {
                "description": "Returns current quote for a ticker symbol (Finnhub)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Get current quote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol (e.g., AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sentiment/{symbol}": {
            "get": {
                "description": "Fetches recent news and returns only the aggregate sentiment values",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sentiment"
                ],
                "summary": "Get aggregate sentiment only",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol (e.g., AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SentimentOnlyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stock/{symbol}": {
            "get": {
                "description": "Returns comprehensive stock detail for a symbol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Get stock detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol (e.g., AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ticker/{symbol}": {
            "get": {
                "description": "Returns aggregated ticker data for the given date (defaults to yesterday)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker"
                ],
                "summary": "Get aggregated ticker data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker symbol (e.g., AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "BertInference.BERTSentiment": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "label": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                }
            }
        },
        "controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "controller.NewsSentimentBERTResponse": {
            "type": "object",
            "properties": {
                "aggregate_result": {
                    "$ref": "#/definitions/sentAnalysis.StockSentimentAnalysis"
                },
                "news_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sentAnalysis.NewsItem"
                    }
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "controller.SentimentOnlyResponse": {
            "type": "object",
            "properties": {
                "sentiment": {
                    "$ref": "#/definitions/sentAnalysis.StockSentimentAnalysis"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "sentAnalysis.NewsItem": {
            "type": "object",
            "properties": {
                "bert_sentiment": {
                    "description": "\u003c-- FinBERT sentiment*",
                    "allOf": [
                        {
                            "$ref": "#/definitions/BertInference.BERTSentiment"
                        }
                    ]
                },
                "content": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "sentiment": {
                    "$ref": "#/definitions/sentAnalysis.Sentiment"
                },
                "symbols": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "sentAnalysis.Sentiment": {
            "type": "object",
            "properties": {
                "neg": {
                    "type": "number"
                },
                "neu": {
                    "type": "number"
                },
                "polarity": {
                    "type": "number"
                },
                "pos": {
                    "type": "number"
                }
            }
        },
        "sentAnalysis.StockSentimentAnalysis": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "negative_ratio": {
                    "type": "number"
                },
                "neutral_ratio": {
                    "type": "number"
                },
                "news_count": {
                    "type": "integer"
                },
                "overall_sentiment": {
                    "description": "-1.0 to 1.0",
                    "type": "number"
                },
                "positive_ratio": {
                    "type": "number"
                },
                "recommendation": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        }
    }
}